#[161912910] Login functionality implemented
