+++ server/src/v1/controllers/parcels.js
@@ -1,73 +1,19 @@
 import Validator from '../helpers/validator';
-
-export const statuses = {
-  AwaitingProcessing: {
-    code: 0,
-    uiText: 'Awaiting Processing',
-  },
-  InTransit: {
-    code: 1,
-    uiText: 'In Transit',
-  },
-  Delivered: {
-    code: 2,
-    uiText: 'Parcel Delivered',
-  },
-  Cancelled: {
-    code: 4,
-    uiText: 'Parcel Delivery Cancelled',
-  },
-};
+import { parcelHelpers, statuses } from '../helpers/mockdb';
 
 const officeLocation = 'Maryland, Lagos';
 const defaultPrice = 'N500';
 
-const Parcels = [
-  {
-    id: 1,
-    destination: 'Ojota, Lagos',
-    pickUpLocation: 'SendIT Pickup Station, Ojota',
-    price: 'N500',
-    status: statuses.InTransit,
-    presentLocation: 'Ketu, Lagos',
-  },
-  {
-    id: 2,
-    destination: 'Ikeja, Lagos',
-    pickUpLocation: 'SendIT Pickup Station, Ikeja',
-    price: 'N700',
-    status: statuses.Delivered,
-  },
-  {
-    id: 3,
-    destination: 'Maitama, Abuja',
-    pickUpLocation: 'SendIT Pickup Station, Wuse',
-    price: 'N5000',
-    status: statuses.Cancelled,
-  },
-];
 
 const ParcelsController = {
-  findAll() {
-    return Parcels || [];
-  },
-
-  find(id) {
-    return Parcels.filter(parcels => parcels.id === id)[0];
-  },
-
-  retrieveParcels(parcelIds) {
-    return Parcels.filter(parcel => parcelIds.indexOf(parcel.id) !== -1);
-  },
-
   index(req, res) {
-    const parcels = ParcelsController.findAll();
+    const parcels = parcelHelpers.findAll();
     return res.json(parcels);
   },
 
   get(req, res) {
     const parcelId = parseInt(req.params.id, 10);
-    const parcel = ParcelsController.find(parcelId);
+    const parcel = parcelHelpers.find(parcelId);
     if (parcel === undefined || parcel === null) {
       return res.status(400).send({ error: 'Parcel delivery order not found.' });
     }
@@ -76,7 +22,7 @@ const ParcelsController = {
 
   cancel(req, res) {
     const parcelId = parseInt(req.params.id, 10);
-    const parcel = ParcelsController.find(parcelId);
+    const parcel = parcelHelpers.find(parcelId);
     if (parcel === undefined || parcel === null) {
       return res.status(400).send({ error: 'Parcel delivery order not found.' });
     }
@@ -87,8 +33,7 @@ const ParcelsController = {
     const newParcel = {};
     Object.assign(newParcel, parcel, { status: statuses.Cancelled });
     delete newParcel.presentLocation;
-    const index = Parcels.indexOf(parcel);
-    Parcels[index] = newParcel;
+    parcelHelpers.update(newParcel);
     return res.json(newParcel);
   },
 
@@ -102,11 +47,11 @@ const ParcelsController = {
         errors,
       });
     }
-    parcel.id = Parcels.length + 1;
+    parcel.id = parcelHelpers.findAll().length + 1;
     parcel.price = defaultPrice;
     parcel.status = statuses.AwaitingProcessing;
     parcel.presentLocation = officeLocation;
-    Parcels.push(parcel);
+    parcelHelpers.addParcel(parcel);
     return res.json(parcel);
   },
 };
