+++ server/src/v1/middlewares/index.js
@@ -2,8 +2,10 @@ import jwt from 'jsonwebtoken';
 
 export default class Middleware {
   static isAuth(req, res, next) {
-    if ((req.url.endsWith('/users/signup') ||req.url.endsWith('/users/signin')) && req.method === 'POST') return next();
-    let token = req.headers['x-access-token'] || req.headers.authorization;
+    let interest = req.url.split('/api/v')[1];
+    interest = interest.substr(2, 4);
+    if ((interest === 'auth') && req.method === 'POST') return next();
+    let token = req.headers.authorization;
     if (token && token.startsWith('Bearer ')) {
       token = token.slice(7, token.length);
     }
