+++ server/src/v1/routes/auth.js
@@ -1,0 +1,7 @@
+import { Router } from 'express';
+import authController from '../controllers/auth';
+
+const router = Router();
+router.post('/signup', authController.signup);
+router.post('/login', authController.login);
+export default router;
